Minileven: add options back  as they still exist on sites (#14756)
* Minileven: add options back  as they still exist on sites

This is a follow-up to #14714.
After removing those options, we started seeing errors like this one on sites:
PHP Warning: Invalid Jetpack option name: wp_mobile_excerpt

Since these Jetpack options are still valid options even if the feature is gone, let's keep them around.

* Those options are not stored as Jetpack options.

See https://github.com/Automattic/jetpack/pull/14756#issuecomment-589860404

* Remove unneeded options

See https://github.com/Automattic/jetpack/pull/14756/files#r383413182

* Only delete options if the standalone plugin is not in use

+++ class.jetpack.php
@@ -435,18 +435,27 @@ class Jetpack {
 					update_option( 'wpcom_publish_posts_with_markdown', true );
 				}
 
-				// Minileven deprecation. 8.3.0.
-				if ( get_option( 'wp_mobile_custom_css' ) ) {
-					delete_option( 'wp_mobile_custom_css' );
-				}
-				if ( Jetpack_Options::get_option( 'wp_mobile_excerpt' ) ) {
-					Jetpack_Options::delete_option( 'wp_mobile_excerpt' );
-				}
-				if ( Jetpack_Options::get_option( 'wp_mobile_featured_images' ) ) {
-					Jetpack_Options::delete_option( 'wp_mobile_featured_images' );
-				}
-				if ( Jetpack_Options::get_option( 'wp_mobile_app_promos' ) ) {
-					Jetpack_Options::delete_option( 'wp_mobile_app_promos' );
+				/*
+				 * Minileven deprecation. 8.3.0.
+				 * Only delete options if not using
+				 * the replacement standalone Minileven plugin.
+				 */
+				if (
+					! self::is_plugin_active( 'minileven-master/minileven.php' )
+					&& ! self::is_plugin_active( 'minileven/minileven.php' )
+				) {
+					if ( get_option( 'wp_mobile_custom_css' ) ) {
+						delete_option( 'wp_mobile_custom_css' );
+					}
+					if ( get_option( 'wp_mobile_excerpt' ) ) {
+						delete_option( 'wp_mobile_excerpt' );
+					}
+					if ( get_option( 'wp_mobile_featured_images' ) ) {
+						delete_option( 'wp_mobile_featured_images' );
+					}
+					if ( get_option( 'wp_mobile_app_promos' ) ) {
+						delete_option( 'wp_mobile_app_promos' );
+					}
 				}
 
 				if ( did_action( 'wp_loaded' ) ) {
